refactor: change organization of traits into single files based on contract
