+++ integration-tests/tests/test_utils/trait_mod/multisig_trait.rs
@@ -1,0 +1,50 @@
+use cosmwasm_std::Addr;
+use cw_multi_test::{App, ContractWrapper, Executor};
+use integration_tests::contract::Contract;
+
+#[derive(Clone)]
+pub struct MultisigContract {
+    pub contract_addr: Addr,
+}
+
+impl MultisigContract {
+    pub fn instantiate_contract(
+        app: &mut App,
+        governance: Addr,
+        rewards_address: Addr,
+        block_expiry: u64,
+    ) -> Self {
+        let code = ContractWrapper::new(
+            multisig::contract::execute,
+            multisig::contract::instantiate,
+            multisig::contract::query,
+        );
+        let code_id = app.store_code(Box::new(code));
+
+        let contract_addr = app
+            .instantiate_contract(
+                code_id,
+                Addr::unchecked("anyone"),
+                &multisig::msg::InstantiateMsg {
+                    rewards_address: rewards_address.to_string(),
+                    governance_address: governance.to_string(),
+                    block_expiry,
+                },
+                &[],
+                "multisig",
+                None,
+            )
+            .unwrap();
+
+        MultisigContract { contract_addr }
+    }
+}
+
+impl Contract for MultisigContract {
+    type QMsg = multisig::msg::QueryMsg;
+    type ExMsg = multisig::msg::ExecuteMsg;
+
+    fn contract_address(&self) -> Addr {
+        self.contract_addr.clone()
+    }
+}
