+++ integration-tests/tests/test_utils/trait_mod/voting_verifier_trait.rs
@@ -1,0 +1,63 @@
+use axelar_wasm_std::nonempty;
+use axelar_wasm_std::MajorityThreshold;
+use connection_router_api::ChainName;
+use cosmwasm_std::Addr;
+use cw_multi_test::{App, ContractWrapper, Executor};
+use integration_tests::contract::Contract;
+
+#[derive(Clone)]
+pub struct VotingVerifierContract {
+    pub contract_addr: Addr,
+}
+
+impl VotingVerifierContract {
+    pub fn instantiate_contract(
+        app: &mut App,
+        service_registry_address: nonempty::String,
+        service_name: nonempty::String,
+        source_gateway_address: nonempty::String,
+        voting_threshold: MajorityThreshold,
+        block_expiry: u64,
+        confirmation_height: u64,
+        source_chain: ChainName,
+        rewards_address: Addr,
+    ) -> Self {
+        let code = ContractWrapper::new(
+            voting_verifier::contract::execute,
+            voting_verifier::contract::instantiate,
+            voting_verifier::contract::query,
+        );
+        let code_id = app.store_code(Box::new(code));
+
+        let contract_addr = app
+            .instantiate_contract(
+                code_id,
+                Addr::unchecked("anyone"),
+                &voting_verifier::msg::InstantiateMsg {
+                    service_registry_address,
+                    service_name,
+                    source_gateway_address,
+                    voting_threshold,
+                    block_expiry,
+                    confirmation_height,
+                    source_chain,
+                    rewards_address: rewards_address.to_string(),
+                },
+                &[],
+                "voting_verifier",
+                None,
+            )
+            .unwrap();
+
+        VotingVerifierContract { contract_addr }
+    }
+}
+
+impl Contract for VotingVerifierContract {
+    type QMsg = voting_verifier::msg::QueryMsg;
+    type ExMsg = voting_verifier::msg::ExecuteMsg;
+
+    fn contract_address(&self) -> Addr {
+        self.contract_addr.clone()
+    }
+}
